Remove PageJoiner interface from non-spilling join operator
Since there is only one implementation - DefaultPageJoiner, which is now called PageJoiner
